Перестроил MySQL интерфейс. Разбил его на части в разных классах
